requirements.txt
Django==1.9.5
Pillow==3.2.0
django-ckeditor==5.0.3
django_extensions==1.4.9